Add authorization controls

+++ app/Http/Controllers/CapacitacionesController.php
@@ -7,6 +7,11 @@ use App;
 
 class CapacitacionesController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,6 +19,7 @@ class CapacitacionesController extends Controller
      */
     public function index()
     {
+        Gate::authorize('admin-stuff');
         $capacitaciones = App\Capacitaciones::paginate(20);
         return view('capacitaciones.index', ['capacitaciones' => $capacitaciones]);
     }
@@ -25,6 +31,7 @@ class CapacitacionesController extends Controller
      */
     public function create()
     {
+        Gate::authorize('admin-stuff');
         $niveles = App\NivelesCapacitaciones::all();
         return view('capacitaciones.create', ['niveles' => $niveles]);
     }
@@ -37,6 +44,7 @@ class CapacitacionesController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('admin-stuff');
         $request->validate([
             'descripcion' => 'required|string',
             'nivel' => 'required|string|exists:niveles_capacitaciones,id',
@@ -62,6 +70,7 @@ class CapacitacionesController extends Controller
      */
     public function edit($id)
     {
+        Gate::authorize('admin-stuff');
         $capacitacione = App\Capacitaciones::findOrFail($id);
         $niveles = App\NivelesCapacitaciones::all();
         return view('capacitaciones.update', [
@@ -79,6 +88,7 @@ class CapacitacionesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('admin-stuff');
         $request->validate([
             'descripcion' => 'required|string',
             'nivel' => 'required|string|exists:niveles_capacitaciones,id',
@@ -105,6 +115,7 @@ class CapacitacionesController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('admin-stuff');
         $capacitacion = App\Capacitaciones::findOrFail($id);
         $capacitacion->delete();
         return redirect()->route('capacitaciones.index');
